optional sympy
