+++ tests/bofire/data_models/test_constraints.py
@@ -192,17 +192,3 @@ def test_constraints_jacobian(constraints, num_candidates):
                 if not hasattr(col, "__iter__"):
                     res[j] = pd.Series(np.repeat(col, candidates.shape[0]))
             assert np.allclose(returned[i], pd.DataFrame(res).transpose())
-
-
-def test_nonlinear_constraints_jacobian_expression():
-    constraint0 = NonlinearInequalityConstraint(
-        expression="x1**2 + x2**2 - x3", features=["x1", "x2", "x3"]
-    )
-    constraint1 = NonlinearInequalityConstraint(
-        expression="x1**2 + x2**2 - x3",
-        features=["x1", "x2", "x3"],
-        jacobian_expression="[2*x1, 2*x2, -1]",
-    )
-
-    data = pd.DataFrame(np.random.rand(10, 3), columns=["x1", "x2", "x3"])
-    assert np.allclose(constraint0.jacobian(data), constraint1.jacobian(data))
